writing more tests and introduced TransactionService implementation

+++ src/main/java/com/resurs/interview/model/api/CustomerResponse.java
@@ -3,7 +3,7 @@ package com.resurs.interview.model.api;
 import com.resurs.interview.db.CustomerEntity;
 
 public record CustomerResponse(Long id, String firstName, String lastName, String email, String address,
-                               String postcode, String city, String socialSecurityNumber, Integer creditScore) {
+                               String postcode, String city, String socialSecurityNumber, Double creditScore) {
 
     public static CustomerResponse transform(CustomerEntity customerEntity) {
         return new CustomerResponse(
